tools/requirements_apt.txt
python3
python3-pip
python3-venv
python3-mysql.connector
mariadb-client
mariadb-server
curl
jq
net-tools
